+++ nlist_mcc/NList-Test.c
@@ -575,7 +575,8 @@ HOOKPROTONHNO(DisplayLI_TextFunc, void, struct NList_DisplayMessage *ndm)
     else if (entry->num % 20 == 13)
       ndm->strings[0] = (STRPTR)"\033o[1]";
     else
-    { snprintf(buf, sizeof(buf), "%ld",entry->num);
+    { 
+      snprintf(buf, sizeof(buf), "%d", (int)entry->num);
       ndm->strings[0]  = buf;
     }
 
@@ -1040,7 +1041,7 @@ int main(UNUSED int argc, UNUSED char *argv[])
     ULONG sigs = 0;
     char *line;
 
-    while ((id = DoMethod(APP_Main,MUIM_Application_NewInput,&sigs)) != MUIV_Application_ReturnID_Quit)
+    while ((id = DoMethod(APP_Main,MUIM_Application_NewInput,&sigs)) != (LONG)MUIV_Application_ReturnID_Quit)
     {
       if (id == ID_LIST2_ACTIVE)
       { DoMethod(LI_Text2, MUIM_NList_GetEntry, MUIV_NList_GetEntry_Active, &line);
